新增查询用户账单信息功能（完成90%）

+++ reckoning/src/main/java/cn/yugutou/reckoning/dto/resp/QueryBillingInfoResp.java
@@ -9,5 +9,5 @@ public class QueryBillingInfoResp implements Serializable {
     //用户账单信息
     private List userBillingInfo;
     //用户账单总条数
-    private Integer totals;
+    private Long totals;
 }
